agent/requirements.txt
livekit-agents>=0.12.16,<0.13.0
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
motor>=3.3.0
pymongo>=4.6.0
pinecone>=3.0.0
sentence-transformers>=2.2.2
certifi